Move FDI ownership transition to mechanisms
