+++ src/main/scala/com/boombustgroup/amorfati/engine/README.md
@@ -108,12 +108,11 @@ population/state transitions, and materializes the month-`t+1` engine boundary.
 | `WorldAssemblyEconomics.scala` | Public `StepInput` / `PostResult` contract and top-level ordering for post-month assembly. |
 | `WorldStateAssembler.scala` | Builds the post-stage `World` value from explicit stage outputs, observables, informal-economy state, and flow-of-funds diagnostics. |
 | `FlowStateAssembler.scala` | Maps stage outputs into `FlowState`, the diagnostic flow surface persisted on `World`. |
-| `PostMonthPopulationTransitions.scala` | Completes post-month population transitions: FDI M&A, firm entry, startup staffing, regional migration, firm-flow birth/death diagnostics, and firm ledger refresh. |
+| `PostMonthPopulationTransitions.scala` | Completes post-month population transitions by invoking domain mechanisms for FDI M&A and firm entry, then applying startup staffing, regional migration, firm-flow birth/death diagnostics, and firm ledger refresh. |
 | `PostMonthPipelineState.scala` | Persists same-month pipeline fields that become visible before next-month signal extraction. |
 | `WorldInformalEconomy.scala` | Computes tax evasion loss, realized shadow-tax share, and smoothed informal-economy cyclical state. |
 | `WorldObservables.scala` | Computes assembled-world observables such as deposit-facility usage, ETS price, and tourism seasonality. |
 | `FlowOfFundsDiagnostics.scala` | Computes the flow-of-funds residual from realized firm revenue and adjusted demand. |
-| `FdiOwnershipTransitions.scala` | Applies post-month stochastic conversion of eligible domestic firms to foreign ownership. |
 | `FirmEntryTransitions.scala` | Derives automation-native entrant diagnostics from newly created firms. |
 | `StartupStaffing.scala` | Assigns startup workers after firm entry and synchronizes startup filled-worker counts with household employment. |
 
@@ -190,6 +189,7 @@ don't clear markets themselves.
 |------|--------|
 | `EuFunds.scala` | EU structural funds: Beta-curve absorption timing, co-financing, capital investment |
 | `Expectations.scala` | Inflation expectations: adaptive-anchoring hybrid, central bank credibility |
+| `FdiOwnershipTransitions.scala` | Stochastic FDI M&A mechanism: eligible domestic firms may become foreign-owned at the post-month transition boundary. |
 | `FirmEntry.scala` | Endogenous firm entry: profit-weighted sector choice, regulatory barriers, AI-native startups |
 | `Macroprudential.scala` | CCyB (countercyclical capital buffer), credit-to-GDP gap, O-SII buffers |
 | `SectoralMobility.scala` | Cross-sector labor transitions: friction matrix, voluntary quits, wage penalties |
