+++ src/main/scala/com/boombustgroup/amorfati/engine/mechanisms/FdiOwnershipTransitions.scala
@@ -1,4 +1,4 @@
-package com.boombustgroup.amorfati.engine.assembly
+package com.boombustgroup.amorfati.engine.mechanisms
 
 import com.boombustgroup.amorfati.agents.Firm
 import com.boombustgroup.amorfati.config.SimParams
